test/web-platform-tests/to-upstream/css/css-backgrounds/background-webkit.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>background shorthand and webkit prefixed property</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/4153 -->

<div id="container"></div>

<script>
"use strict";

test(() => {
  const container = document.getElementById("container");
  const styleText = "background: red; background-clip: padding-box; -webkit-text-fill-color: transparent;";
  container.innerHTML = `<div id="div" style="${styleText}"></div>`;
  const div = document.getElementById("div");
  assert_equals(div.style.cssText, "background: padding-box red; -webkit-text-fill-color: transparent;");
}, "putting -webkit property value after background longhand");
</script>
